- Temporarily deactivated all the UseScroll and Drinking stuff as it seems that the code crashing the client is the one dealing with items in bags. - Small fixes to the way percentages are calculated in the wrapper (dealing with int that should have been float)

+++ BabBot/BabBot/Scripting/PlayerWrapper.cs
@@ -266,12 +266,12 @@ namespace BabBot.Scripting
 
         public int HpPct()
         {
-            return (int)((player.Hp/player.MaxHp)*100);
+            return (int)((float)((float)player.Hp / (float)player.MaxHp) * 100);
         }
 
         public int MpPct()
         {
-            return (int)((player.Mp / player.MaxMp) * 100);
+            return (int)((float)((float)player.Mp / (float)player.MaxMp) * 100);
         }
 
         public bool IsTargetDead()
